refactor: cn 함수를 별도 파일로 분리
- src/shared/lib/cn.ts 생성 (shadcn/ui cn 함수)
- src/shared/lib/utils.ts에서 cn 제거 (formatDate, formatTimeAgo 유지)
- src/lib/utils.ts 삭제 (shadcn 기본 위치 제거)
- src/shared/ui 컴포넌트들 import 경로 변경 (@/shared/lib/cn)
- 테스트 파일 분리
  - tests/unit/shared/lib/cn.test.ts 생성
  - tests/unit/shared/lib/utils.test.ts에서 cn 테스트 제거
