+++ tests/unit/shared/lib/utils.test.ts
@@ -1,19 +1,6 @@
-import { cn, formatDate, formatTimeAgo } from "@/shared/lib/utils";
+import { formatDate, formatTimeAgo } from "@/shared/lib/utils";
 
 describe("utils", () => {
-  describe("cn", () => {
-    it("should merge class names correctly", () => {
-      expect(cn("px-2 py-1", "px-4")).toBe("py-1 px-4");
-    });
-
-    it("should handle conditional classes", () => {
-      expect(cn("base-class", true && "conditional-class")).toBe(
-        "base-class conditional-class"
-      );
-      expect(cn("base-class", false && "conditional-class")).toBe("base-class");
-    });
-  });
-
   describe("formatDate", () => {
     it("should format Date object correctly", () => {
       const date = new Date("2024-01-15");
